[nxp noup] platform: rdrw61x: Add patch for enabling rollback protection for ITS.
For rollback protection in ITS, the approach is to use two additional nv
counters. Definition of IDs of nv_counters as well as checking of usage
permissions for those counters is done in TF-M core code and there is no
way for a platform to extend this easily.

Also, as the nv counters are implemented as part of the platform service
and get used from the ITS service, a dependency from ITS on the platform
service is required (in a similar fashion as it is required for the
rollback protection of PS).

As this requires platform specific changes to TF-M core code, this is
included here as a patch.

Signed-off-by: Stefan Krug <stefan.krug@nxp.com>